AbyssalCraft 1.9.0- The Omothol Update
*Added missing page to the Potential Energy section of Ritual
Information (mentions how monoliths are created)
*Darklands Villages should now have Darklands Oak Pressure Plates and
Darkstone Double slabs (instead of the vanilla equivalents)
*Depths Ghouls, Abyssal Zombies and Lesser Shoggoths should no longer
spawn at Mushroom Island Shores
*Fixed some Necronomicon Pages with the wrong text
*Added a config option that allows you to change how Evil Animals spawn
*Disabled any test code in the Materializer that could trigger a crash
(since it's incomplete)
*Lesser Shoggoths outside of the Overworld should drop the correct item
when killed now (applying metadata to the dropped item)
*Spectral Dragons should now be easier to hit without getting hit from
them
*Spectral Dragon spawn rate has been reduced to the lowest possible
(encountering one now should be very rare, unless another mod tweaks the
spawn rate)
*Darklands Oak Slabs are no longer treated as stone
*Added a config option allowing you to toggle whether or not Liquid
Antimatter should disintegrate any item dropped into it
*Replaced the Fist of Cha'garoth spawner block with a J'zahar spawner
block
*J'zahar has increased in size
*If a Spectral Dragon dies while Asorah is draining it, Asorah won't
take explosion damage
*Any item placed on a Ritual/Sacrificial Altar or Ritual/Energy Pedestal
will also render for other players
*The smoke projected when a ritual is performed also renders for other
players
*Shoggoth Lairs should now longer spawn in biomes registered to the
BiomeDictionary as the "WATER" type
*The Staff of Rending will now continue to drain energy when the right
mouse button is pressed
*Shadow Fragments/Shards/Gems now spawn in
dungeon/mineshaft/pyramid/stronghold chests
*The Shard of Oblivion now requires 4 Shadow Gems (previously 8)
*Remnants can now open doors
*Added the Temple of J'zahar (generates in Omothol)
*Added the City of Omothol
*Updated the Omothol Progression section of the Necronomicon
*Added Essence of The Gatekeeper (replacement for the Oblivion Catalyst
in the Abyssalnomicon recipe, dropped by J'zahar)
*Staff of The Gatekeeper is now created through a ritual (instead of
dropped by J'zahar)
*Added a Minion of The Gatekeeper spawner block
*Remnants and Minions of The Gatekeeper no longer despawn
*Spawn rate of Remnants and Minions of The Gatekeeper has been heavily
reduced (quite rare now)
*Minions of The Gatekeeper now has a 10% chance of dropping Ethaxium
Ingots (while Eldritch Scales are the regular drop)

+++ src/main/java/com/shinoow/abyssalcraft/common/disruptions/DisruptionFireRain.java
@@ -77,11 +77,11 @@ public class DisruptionFireRain extends DisruptionEntry {
 		EntityLargeFireball fireball = new EntityLargeFireball(world);
 		fireball.setLocationAndAngles(x, y, z, fireball.rotationYaw, fireball.rotationPitch);
 		fireball.setPosition(x, y, z);
-		double d6 = (double)MathHelper.sqrt_double(accx * accx + accy * accy + accz * accz);
+		double d6 = MathHelper.sqrt_double(accx * accx + accy * accy + accz * accz);
 		fireball.accelerationX = accx / d6 * 0.1D;
 		fireball.accelerationY = accy / d6 * 0.1D;
 		fireball.accelerationZ = accz / d6 * 0.1D;
-		
+
 		return fireball;
 	}
 }
